* Moved Rabi to characterization.
