+++ qiskit_experiments/library/__init__.py
@@ -65,6 +65,8 @@ Some experiments may be also used for gate calibration.
     ~characterization.FineAmplitude
     ~characterization.FineXAmplitude
     ~characterization.FineSXAmplitude
+    ~characterization.Rabi
+    ~characterization.EFRabi
 
 .. _calibration:
 
@@ -86,8 +88,6 @@ See :doc:`/tutorials/calibrating_armonk` for example.
     ~calibration.FineDrag
     ~calibration.FineXDrag
     ~calibration.FineSXDrag
-    ~calibration.Rabi
-    ~calibration.EFRabi
     ~calibration.FineAmplitudeCal
     ~calibration.FineXAmplitudeCal
     ~calibration.FineSXAmplitudeCal
@@ -99,8 +99,6 @@ from .calibration import (
     FineDrag,
     FineXDrag,
     FineSXDrag,
-    Rabi,
-    EFRabi,
     FineAmplitudeCal,
     FineXAmplitudeCal,
     FineSXAmplitudeCal,
@@ -114,6 +112,8 @@ from .characterization import (
     EFSpectroscopy,
     CrossResonanceHamiltonian,
     EchoedCrossResonanceHamiltonian,
+    Rabi,
+    EFRabi,
     FineAmplitude,
     FineXAmplitude,
     FineSXAmplitude,
