+++ qiskit_experiments/library/characterization/__init__.py
@@ -28,6 +28,8 @@ Experiments
     QubitSpectroscopy
     CrossResonanceHamiltonian
     EchoedCrossResonanceHamiltonian
+    Rabi
+    EFRabi
     FineAmplitude
     FineXAmplitude
     FineSXAmplitude
@@ -54,4 +56,5 @@ from .t2ramsey import T2Ramsey
 from .t2ramsey_analysis import T2RamseyAnalysis
 from .cr_hamiltonian import CrossResonanceHamiltonian, EchoedCrossResonanceHamiltonian
 from .cr_hamiltonian_analysis import CrossResonanceHamiltonianAnalysis
+from .rabi import Rabi, EFRabi
 from .fine_amplitude import FineAmplitude, FineXAmplitude, FineSXAmplitude
